+++ src/app/pages/settings/settings.component.ts
@@ -311,7 +311,7 @@ export class SettingsPageComponent {
   }
 
   getAppVersion() {
-    return environment.version || '1.0.8';
+    return environment.version || '1.0.9';
   }
 
   getPlatform() {