+++ src/environments/environment.prod.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
-  version: '1.0.8',
+  version: '1.0.9',
   curseForgeApiKey: 'CURSEFORGE_KEY_PLACEHOLDER'
 };